Fold the grid-array readers into gridState.js
Move get32NoteArrayFromClickableUI and muteArrayFromClickableUI (plus the
pure get_empty_note_array / fill_array_with_value_false helpers) out of
GrooveWriter into gridState.js, so the module is now the complete DOM
grid-read layer: cell state through whole-measure arrays. The readers call
the note-state getters directly and import the pure getNoteScaler; their
GrooveWriter-owned deps are injected (layout + row-visibility via a ctx
object, isInstrumentMuted as a callback) rather than pulled in. GrooveWriter
keeps same-name wrappers, so every call site is unchanged.

Behavior-neutral: 697 unit + 71 E2E golden-master pass (SVG/MIDI
byte-identical); typecheck neutral, 0 lint errors, knip clean.
